Use `--not` twice in git aliases
`--not` changes the interpretation of all revisions that appear to
the right of it. A second `--not` will switch back to the default
interpretation.

So `--not A` isn't actually equivalent to `^A`. Rather, `--not A`
is equivalent to `^A` **AND** adding `^` to every following
revision.

`--not A --not` is actually equivalent to `^A`, so use that in
aliases.